Edited node equals and hashcode

+++ src/main/java/cz/certicon/routing/model/entity/Edge.java
@@ -153,7 +153,7 @@ public interface Edge {
         }
 
         public static Id fromString( String str ) {
-            return createId( Integer.parseInt( str ) );
+            return createId( Long.parseLong( str ) );
         }
 
         private final long id;
